+++ src/components/Form Fields/error-msg.tsx
@@ -4,8 +4,10 @@ import {
   UseControllerProps
 } from "react-hook-form";
 
-export type FormProps<T extends FieldValues = FieldValues> =
-  UseControllerProps<T>;
+export interface FormProps<T extends FieldValues = FieldValues>
+  extends UseControllerProps<T> {
+  className?: string;
+}
 
 /*
 This is a component to display the error messages for an input field
@@ -13,13 +15,20 @@ Output:
   A <div> that contains the error message for a specific input field
 */
 export default function Error<T extends FieldValues = FieldValues>({
+  className,
   ...props
 }: FormProps<T>) {
   const { field, fieldState } = useController(props);
   return (
     <div>
       {/* Display error message */}
-      <p className="absolute right-2 top-6 self-start rounded bg-white px-2 py-1 text-xs italic text-red-500">
+      <p
+        className={
+          !className
+            ? "absolute -bottom-3 right-2 self-start rounded bg-white px-2 py-1 text-xs italic text-red-500"
+            : className
+        }
+      >
         {!fieldState.error?.message
           ? `${field.name} is invalid`
           : fieldState.error.message}
